Move from integration to integrationpoint
